vector-store: extract primary_id mod from table
